Created DropdownUtility component and fixed useEffect bug in GearForm
