+++ src/components/forms/EditGear.jsx
@@ -2,15 +2,10 @@ import { useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { editGearPost, getPostById } from "../services/postService";
 import { GearForm } from "./GearForm";
-import { getCategories } from "../services/categoryService";
-import { getConditions } from "../services/conditionService";
-import { getBrands } from "../services/brandService";
+import { DropdownUtility } from "../filter/DropdownUtility";
 
 export const EditGear = ({ currentUser }) => {
   const [initialData, setInitialData] = useState(null);
-  const [categories, setCategories] = useState([]);
-  const [conditions, setConditions] = useState([]);
-  const [brands, setBrands] = useState([]);
   const { postId } = useParams();
   const navigate = useNavigate();
 
@@ -19,22 +14,6 @@ export const EditGear = ({ currentUser }) => {
       setInitialData(post);
     });
   }, [postId]);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const fetchedCategories = await getCategories();
-        setCategories(fetchedCategories);
-        const fetchedConditions = await getConditions();
-        setConditions(fetchedConditions);
-        const fetchedBrands = await getBrands();
-        setBrands(fetchedBrands);
-      } catch (error) {
-        console.error("Error fetching dropdown data:", error);
-      }
-    };
-    fetchData();
-  }, []);
 
   const handleUpdatePost = async (updatedPost) => {
     const editedPost = {
@@ -53,16 +32,22 @@ export const EditGear = ({ currentUser }) => {
     });
   };
 
-  return initialData ? (
-    <GearForm
-      currentUser={currentUser}
-      initialData={initialData}
-      categories={categories}
-      conditions={conditions}
-      brands={brands}
-      handleSubmit={handleUpdatePost}
-    />
-  ) : (
-    <div>Loading...</div>
+  return (
+    <DropdownUtility>
+      {({ categories, conditions, brands }) =>
+        initialData ? (
+          <GearForm
+            currentUser={currentUser}
+            initialData={initialData}
+            categories={categories}
+            conditions={conditions}
+            brands={brands}
+            handleSubmit={handleUpdatePost}
+          />
+        ) : (
+          <div>Loading...</div>
+        )
+      }
+    </DropdownUtility>
   );
 };
